requirements.txt
requests
beautifulsoup4
gspread
google-auth
selenium
webdriver-manager
